+++ lib/zxing_scanner.dart
@@ -1,6 +1,7 @@
 library zxing_scanner;
 
 export 'src/scan_view.dart';
+export 'src/scan_view_web.dart';
 export 'src/controller.dart';
 export 'src/foundation/utils.dart';