+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* auto compress image
+* camera stream scan & camera picture scan
+
 ## 0.1.0
 
 * supported scan from camera